Nova rotina de gravacao de recebimento

+++ app/controllers/application_controller.rb
@@ -79,6 +79,7 @@ class ApplicationController < ActionController::Base
     end
     
     def verifica_horario_de_trabalho
+      return true if current_user.master?
       if !current_user.horario_de_trabalho?
         current_user_session.destroy
         flash[:notice] = "Volte no seu horário de trabalho !"
